+++ examples/benchmark.rs
@@ -6,13 +6,13 @@ use fnv::FnvHashMap;
 use rmf_reservations::algorithms::greedy_solver::{
     generate_test_scenario_with_known_best, TimeBasedBranchAndBound,
 };
-use rmf_reservations::algorithms::sat::SATSolver;
+use rmf_reservations::algorithms::sat::{SATSolver, generate_sat_devil};
 
 fn main() {
-    for x in 7..16 {
+    for x in 6..10 {
         for _ in 0..100 {
-            let (requests, resources) = //generate_sat_devil(5,3);
-        generate_test_scenario_with_known_best(5, 10, x);
+            let (requests, resources) = generate_sat_devil(x,x-2);
+        //generate_test_scenario_with_known_best(5, 10, x);
             //println!("Requests {:?}", requests);
 
             let mut system = TimeBasedBranchAndBound::create_with_resources(&resources);
@@ -26,7 +26,7 @@ fn main() {
             let optimality_proof_dur = timer.elapsed();
 
             let timer = SystemTime::now();
-            SATSolver::from_hill_climber(soln.clone());
+            //SATSolver::from_hill_climber(soln.clone());
             let brute_force_proof_dur = timer.elapsed();
 
             let hint = FnvHashMap::default();
