+++ src/algorithms/kuhn_munkres.rs
@@ -95,7 +95,7 @@ fn test_sparse_axis_masker() {
 /// In this reservation system, time is ignored,
 /// rather the best assignments between resources. An example use case for this
 /// is in a fire emergency system where for instance robots.
-/// This class also comes with a masker that allows you to add simple constraints whild
+/// This class also comes with a masker that allows you to add simple constraints while
 /// solving an optimization.
 struct ReservationsKuhnMunkres {
     resources: Vec<String>,
